Compute basic values indices as measured in European Social Survey

Computed in line with Shalom Schwartz's instructions on the ESS EduNet website.

Usage

ess_values(
  data,
  v2 = TRUE,
  v4 = TRUE,
  v10 = TRUE,
  v21 = FALSE,
  center = TRUE,
  abbr = FALSE,
  suffix = ""
)

Arguments

data

ESS dataset of any other dataset with the same 21 value indicator variable names

v2

Two higher value dimensions

v4

Four higher order values

v10

Compute 10 value indices

v21

Add 21 raw items

center

Apply within-individual centering?

abbr

Logical. Use abbreviated names, e.g. SE, ST, CO..

suffix

NULL of character. What is added to the names of centered indices.

Value

Returns a data with appended variables.