Make a quick network graph

Make a quick network graph

Usage

get_net(
  df = NULL,
  cv = NULL,
  n = NULL,
  title = "Model",
  legend = "none",
  gamma = 0.3,
  layout = "fr",
  start = "glasso",
  stepwise = T,
  seed = 1,
  group.id = NA
)

Arguments

df

Data frame

cv

Correlation matrix (if df is NULL)

title

Title of the plot

legend

Legend position

gamma

Gamma parameter for the model selection function, see `ggmModSelect`

layout

Default is "fr". Network layout. See more at `ggraph::ggraph`

start

Default is "glasso", see `ggmModSelect`

stepwise

Default is TRUE, see `ggmModSelect`

seed

Attempt to make plots replicable.

group.id

Vector of regex expressions to apply to variable names and use it to group variables.

Details

Can be used within `mutate()` and passed to `knitr::kable` with parameter `kable.options=list(escape=FALSE)`