Conditional formatting

Adds formatting to a vector of parameters

Usage

sig_to_bold(
  x,
  star = F,
  bold.thresh = 1,
  neg.red = T,
  small.thresh = 0.01,
  replace.small = NA
)

Arguments

x

Character vector

star

Removes stars if FALSE.

bold.thresh

Bold font if the absolute value is larger than this threshold.

neg.red

Red font if the value is negative value.

small.thresh

Replaces the value with `<|.01|` if the magnitude is smaller than the indicated number. Default is .01.

Details

Can be used within `mutate()` and passed to `knitr::kable` with parameter `kable.options=list(escape=FALSE)`